+++ src/configs/FamConfig.js
@@ -889,7 +889,7 @@ export default function FamConfig(name) {
    * @group Level Titles
    * @type {TextOrientationType}
    */
-  this.levelTitleOrientation = TextOrientationType.RotateRight;
+  this.levelTitleOrientation = TextOrientationType.Auto;
 
   /**
    * Level titles vertical alignment.